+++ src/commands/cli-101-training/feedback.js
@@ -1,4 +1,4 @@
-const { Command } = require('@oclif/command');
+const { Command } = require('@oclif/core');
 
 class Feedback extends Command {
   async run() {